♻️ (#39) refactor: 설정 페이지를 회원 설정, 가게 설정으로 분리
