+++ src/features/account-settings/components/sections/AccountSection.tsx
@@ -1,10 +1,10 @@
-import { KeyRound, LogOut, UserCircle } from 'lucide-react';
+import { LogOut, UserCircle } from 'lucide-react';
 
-import SettingRow from '@/features/settings/components/SettingRow';
-import SettingsSection from '@/features/settings/components/SettingsSection';
-import { MOCK_ACCOUNT_SETTINGS } from '@/features/settings/data/settings.mock';
+import { MOCK_ACCOUNT_SETTINGS } from '@/features/account-settings/data/account-settings.mock';
 import { Button } from '@/shadcn/ui/button';
 import { Separator } from '@/shadcn/ui/separator';
+import SettingRow from '@/shared/components/SettingRow';
+import SettingsSection from '@/shared/components/SettingsSection';
 
 const AccountSection = () => {
   return (
@@ -16,24 +16,6 @@ const AccountSection = () => {
       <SettingRow
         label="이름"
         action={<span className="text-sm text-muted-foreground">{MOCK_ACCOUNT_SETTINGS.name}</span>}
-      />
-      <Separator />
-      <SettingRow
-        label="이메일"
-        action={
-          <span className="text-sm text-muted-foreground">{MOCK_ACCOUNT_SETTINGS.email}</span>
-        }
-      />
-      <Separator />
-      <SettingRow
-        label="비밀번호"
-        description="주기적으로 비밀번호를 변경하면 계정을 안전하게 보호할 수 있습니다."
-        action={
-          <Button variant="outline" size="sm" className="gap-1.5">
-            <KeyRound className="h-3.5 w-3.5" />
-            변경
-          </Button>
-        }
       />
       <Separator />
       <SettingRow
