+++ src/features/account-settings/components/sections/NotificationSection.tsx
@@ -2,12 +2,12 @@ import { useState } from 'react';
 
 import { Bell } from 'lucide-react';
 
-import SettingRow from '@/features/settings/components/SettingRow';
-import SettingsSection from '@/features/settings/components/SettingsSection';
-import { MOCK_NOTIFICATION_SETTINGS } from '@/features/settings/data/settings.mock';
-import type { NotificationSettings } from '@/features/settings/types/settings.types';
+import { MOCK_NOTIFICATION_SETTINGS } from '@/features/account-settings/data/account-settings.mock';
+import type { NotificationSettings } from '@/features/account-settings/types/account-settings.types';
 import { Separator } from '@/shadcn/ui/separator';
 import { Switch } from '@/shadcn/ui/switch';
+import SettingRow from '@/shared/components/SettingRow';
+import SettingsSection from '@/shared/components/SettingsSection';
 
 const NotificationSection = () => {
   const [settings, setSettings] = useState<NotificationSettings>(MOCK_NOTIFICATION_SETTINGS);
@@ -15,13 +15,6 @@ const NotificationSection = () => {
   const toggleAlert = (key: keyof Omit<NotificationSettings, 'channels'>) => {
     setSettings((s) => ({ ...s, [key]: !s[key] }));
   };
-
-  // const toggleChannel = (key: keyof NotificationSettings['channels']) => {
-  //   setSettings((s) => ({
-  //     ...s,
-  //     channels: { ...s.channels, [key]: !s.channels[key] },
-  //   }));
-  // };
 
   return (
     <SettingsSection
@@ -72,43 +65,6 @@ const NotificationSection = () => {
           />
         }
       />
-
-      {/* <div className="flex flex-col gap-4 pt-2">
-        <p className="text-xs font-medium uppercase tracking-wide text-muted-foreground">
-          수신 채널
-        </p>
-        <div className="space-y-4">
-          <SettingRow
-            label="푸시 알림"
-            action={
-              <Switch
-                checked={settings.channels.push}
-                onCheckedChange={() => toggleChannel('push')}
-              />
-            }
-          />
-          <Separator />
-          <SettingRow
-            label="이메일"
-            action={
-              <Switch
-                checked={settings.channels.email}
-                onCheckedChange={() => toggleChannel('email')}
-              />
-            }
-          />
-          <Separator />
-          <SettingRow
-            label="카카오톡"
-            action={
-              <Switch
-                checked={settings.channels.kakao}
-                onCheckedChange={() => toggleChannel('kakao')}
-              />
-            }
-          />
-        </div>
-      </div> */}
     </SettingsSection>
   );
 };
