+++ src/features/account-settings/data/account-settings.mock.ts
@@ -1,15 +1,7 @@
 import type {
   AccountSettings,
   NotificationSettings,
-  StoreSettings,
-} from '@/features/settings/types/settings.types';
-
-export const MOCK_STORE_SETTINGS: StoreSettings = {
-  storeName: '임포트 감자 식당',
-  ownerName: '유다연',
-  businessType: 'individual',
-  category: 'korean',
-};
+} from '@/features/account-settings/types/account-settings.types';
 
 export const MOCK_ACCOUNT_SETTINGS: AccountSettings = {
   name: '유다연',
