+++ src/features/account-settings/types/account-settings.types.ts
@@ -1,10 +1,3 @@
-export interface StoreSettings {
-  storeName: string;
-  ownerName: string;
-  businessType: 'franchise' | 'directly-operated' | 'individual';
-  category: string;
-}
-
 export interface AccountSettings {
   name: string;
   email: string;
