+++ src/features/store-settings/data/store-settings.mock.ts
@@ -1,0 +1,8 @@
+import type { StoreSettings } from '@/features/store-settings/types/store-settings.types';
+
+export const MOCK_STORE_SETTINGS: StoreSettings = {
+  storeName: '임포트 감자 카페',
+  ownerName: '유다연',
+  businessType: 'individual',
+  category: 'cafe',
+};
