+++ src/features/store-settings/types/store-settings.types.ts
@@ -1,0 +1,6 @@
+export interface StoreSettings {
+  storeName: string;
+  ownerName: string;
+  businessType: 'franchise' | 'directly-operated' | 'individual';
+  category: string;
+}
